+++ Backend/erpc/src/main/java/com/cineverse/erpc/contract/dto/ContractDTO.java
@@ -4,9 +4,8 @@ import com.cineverse.erpc.account.account.aggregate.Account;
 import com.cineverse.erpc.contract.aggregate.ContractCategory;
 import com.cineverse.erpc.contract.aggregate.ContractProduct;
 import com.cineverse.erpc.employee.aggregate.Employee;
-import com.cineverse.erpc.quotation.quotation.aggregate.QuotationProduct;
 import com.cineverse.erpc.quotation.quotation.aggregate.Transaction;
-import com.cineverse.erpc.warehouse.aggregate.entity.Warehouse;
+import com.cineverse.erpc.warehouse.aggregate.Warehouse;
 import lombok.*;
 
 import java.util.List;
