+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/board/repository/NoticeBoardRepository.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.notice.board.repository;
 
-import com.cineverse.erpc.notice.board.aggregate.entity.NoticeBoard;
+import com.cineverse.erpc.notice.board.aggregate.NoticeBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
