+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/comment/repository/NoticeCommentRepository.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.notice.comment.repository;
 
-import com.cineverse.erpc.notice.comment.aggregate.entity.NoticeComment;
+import com.cineverse.erpc.notice.comment.aggregate.NoticeComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
