+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/comment/service/NoticeCommentService.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.notice.comment.service;
 
-import com.cineverse.erpc.notice.comment.aggregate.entity.NoticeComment;
+import com.cineverse.erpc.notice.comment.aggregate.NoticeComment;
 import com.cineverse.erpc.notice.comment.dto.NoticeCommentDTO;
 
 import java.util.List;
