+++ Backend/erpc/src/main/java/com/cineverse/erpc/notice/comment/service/NoticeCommentServiceImpl.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.notice.comment.service;
 
-import com.cineverse.erpc.notice.comment.aggregate.entity.NoticeComment;
+import com.cineverse.erpc.notice.comment.aggregate.NoticeComment;
 import com.cineverse.erpc.notice.comment.dto.NoticeCommentDTO;
 import com.cineverse.erpc.notice.comment.repository.NoticeCommentRepository;
 import jakarta.persistence.EntityNotFoundException;
