+++ Backend/erpc/src/main/java/com/cineverse/erpc/product/aggregate/Product.java
@@ -1,4 +1,4 @@
-package com.cineverse.erpc.product.aggregate.entity;
+package com.cineverse.erpc.product.aggregate;
 
 import jakarta.persistence.*;
 import lombok.*;
