+++ Backend/erpc/src/main/java/com/cineverse/erpc/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.product.repository;
 
-import com.cineverse.erpc.product.aggregate.entity.Product;
+import com.cineverse.erpc.product.aggregate.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
