+++ Backend/erpc/src/main/java/com/cineverse/erpc/product/service/ProductService.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.product.service;
 
-import com.cineverse.erpc.product.aggregate.entity.Product;
+import com.cineverse.erpc.product.aggregate.Product;
 import com.cineverse.erpc.product.dto.ProductDTO;
 
 import java.util.List;
