+++ Backend/erpc/src/main/java/com/cineverse/erpc/quotation/quotation/aggregate/Quotation.java
@@ -2,11 +2,10 @@ package com.cineverse.erpc.quotation.quotation.aggregate;
 
 import com.cineverse.erpc.account.account.aggregate.Account;
 import com.cineverse.erpc.employee.aggregate.Employee;
-import com.cineverse.erpc.warehouse.aggregate.entity.Warehouse;
+import com.cineverse.erpc.warehouse.aggregate.Warehouse;
 import jakarta.persistence.*;
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
