+++ Backend/erpc/src/main/java/com/cineverse/erpc/quotation/quotation/dto/QuotationDTO.java
@@ -3,7 +3,7 @@ package com.cineverse.erpc.quotation.quotation.dto;
 import com.cineverse.erpc.account.account.aggregate.Account;
 import com.cineverse.erpc.employee.aggregate.Employee;
 import com.cineverse.erpc.quotation.quotation.aggregate.Transaction;
-import com.cineverse.erpc.warehouse.aggregate.entity.Warehouse;
+import com.cineverse.erpc.warehouse.aggregate.Warehouse;
 import lombok.Data;
 
 @Data
