+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/note/service/SalesOppNoteService.java
@@ -1,7 +1,7 @@
 package com.cineverse.erpc.salesopp.note.service;
 
 import com.cineverse.erpc.salesopp.note.dto.SalesOppNoteDTO;
-import com.cineverse.erpc.salesopp.note.aggregate.entity.SalesOppNote;
+import com.cineverse.erpc.salesopp.note.aggregate.SalesOppNote;
 
 import java.util.List;
