+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/aggregate/SalesOppStatus.java
@@ -1,4 +1,4 @@
-package com.cineverse.erpc.salesopp.opportunity.aggregate.entity;
+package com.cineverse.erpc.salesopp.opportunity.aggregate;
 
 import jakarta.persistence.*;
 import lombok.*;
