+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/dto/SalesOppDTO.java
@@ -1,7 +1,7 @@
 package com.cineverse.erpc.salesopp.opportunity.dto;
 
 import com.cineverse.erpc.employee.aggregate.Employee;
-import com.cineverse.erpc.salesopp.opportunity.aggregate.entity.SalesOppStatus;
+import com.cineverse.erpc.salesopp.opportunity.aggregate.SalesOppStatus;
 import lombok.*;
 
 @AllArgsConstructor
