+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/repository/SalesOppRepository.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.salesopp.opportunity.repository;
 
-import com.cineverse.erpc.salesopp.opportunity.aggregate.entity.SalesOpp;
+import com.cineverse.erpc.salesopp.opportunity.aggregate.SalesOpp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
