+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/service/SalesOppService.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.salesopp.opportunity.service;
 
-import com.cineverse.erpc.salesopp.opportunity.aggregate.entity.SalesOpp;
+import com.cineverse.erpc.salesopp.opportunity.aggregate.SalesOpp;
 import com.cineverse.erpc.salesopp.opportunity.dto.SalesOppDTO;
 
 import java.util.List;
