+++ Backend/erpc/src/main/java/com/cineverse/erpc/warehouse/repository/WarehouseRepository.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.warehouse.repository;
 
-import com.cineverse.erpc.warehouse.aggregate.entity.Warehouse;
+import com.cineverse.erpc.warehouse.aggregate.Warehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
