+++ Backend/erpc/src/main/java/com/cineverse/erpc/warehouse/service/WarehouseService.java
@@ -1,6 +1,6 @@
 package com.cineverse.erpc.warehouse.service;
 
-import com.cineverse.erpc.warehouse.aggregate.entity.Warehouse;
+import com.cineverse.erpc.warehouse.aggregate.Warehouse;
 
 import java.util.List;
